backend/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-jose>=3.3.0
python-multipart>=0.0.5
passlib[bcrypt]
sqlalchemy>=2.0.4
httpx>=0.23.3
pytest>=7.2.1
psycopg2-binary==2.9.5
pandas>=0.25
seaborn>=0.12.2
requests>=2.28.2
plotly>=5.14.1
python-dotenv>=1.0.1
